+++ android/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">1UP</string>
+    <string name="app_name">ImageProject</string>
 </resources>
