+++ package.json
@@ -7,9 +7,8 @@
   },
   "dependencies": {
     "firebase": "^3.1.0",
-    "react": "^15.1.0",
-    "react-native": "^0.27.2",
-    "react-native-loading-spinner-overlay": "^0.1.2",
-    "react-native-scrollable-tab-view": "^0.5.2"
+    "react": "^15.4.2",
+    "react-native": "^0.40.0",
+    "react-native-scrollable-tab-view": "^0.5.5"
   }
 }
